Files reorganization and comments
