+++ sources/packages/backend/workflow/test/test-utils/factories/assessment-consolidated-data.ts
@@ -5,7 +5,6 @@ import {
   OfferingDeliveryOptions,
 } from "@sims/test-utils";
 import { OfferingIntensity } from "@sims/sims-db";
-import * as faker from "faker";
 
 /**
  * Create fake consolidated data
