+++ sources/packages/backend/workflow/test/test-utils/index.ts
@@ -3,8 +3,3 @@ export * from "./mock/zeebe-worker-mock";
 export * from "./constants/workflow-variables-constants";
 export * from "./constants/system-configurations-constants";
 export * from "./validation/workflow-validation.utils";
-export * from "./mock/create-income-request-task-mock";
-export * from "./mock/verify-application-exceptions-task-mock";
-export * from "./mock/check-supporting-user-response-task-mock";
-export * from "./mock/create-supporting-users-parents-task-mock";
-export * from "./mock/check-income-request-task-mock";
