+++ sources/packages/backend/workflow/test/test-utils/mock/assessment-gateway/create-supporting-users-parents-task-mock.ts
@@ -1,6 +1,14 @@
 import { Duration } from "zeebe-node";
-import { createMockedWorkerResult, WorkflowServiceTasks } from "..";
+import { createMockedWorkerResult, WorkflowServiceTasks } from "../..";
 
+/**
+ * Create the mock for 'Create supporting users for parent/parents' completed task
+ * and publish the necessary messages to unblock the workflow.
+ * @param supportingUserIds ids (1 or 2) of the mocked support users that will
+ * also be created using mocks.
+ * @returns mocked ids and one mocked message for each provided id to unblock
+ * the workflow.
+ */
 export function createCreateSupportingUsersParentsTaskMock(options: {
   supportingUserIds: number[];
 }): Record<string, unknown> {
