+++ sources/packages/backend/workflow/test/test-utils/mock/assessment-gateway/verify-application-exceptions-task-mock.ts
@@ -1,0 +1,23 @@
+import { ApplicationExceptionStatus } from "@sims/sims-db";
+import { createMockedWorkerResult, WorkflowServiceTasks } from "../..";
+
+/**
+ * Create the mock for 'Verify Application Exceptions' completed task.
+ * @param options mock options.
+ * - `subprocesses` subprocess reference when the workflow was invoked
+ * using a call activity. It can be define as one or multiple ones case
+ * the workflow was invoked from subprocess inside another subprocess.
+ * @returns mock for 'Verify Application Exceptions' completed task.
+ */
+export function createVerifyApplicationExceptionsTaskMock(options?: {
+  status: ApplicationExceptionStatus;
+}) {
+  const applicationExceptionStatus =
+    options?.status ?? ApplicationExceptionStatus.Approved;
+  return createMockedWorkerResult(
+    WorkflowServiceTasks.VerifyApplicationExceptions,
+    {
+      jobCompleteMock: { applicationExceptionStatus },
+    },
+  );
+}
