+++ sources/packages/backend/workflow/test/test-utils/mock/cra-integration-income-verification/check-income-request-task-mock.ts
@@ -2,8 +2,16 @@ import {
   createMockedWorkerResult,
   WorkflowServiceTasks,
   WorkflowSubprocesses,
-} from "..";
+} from "../..";
 
+/**
+ * Create the mock for 'Check income request' task.
+ * @param options mock options.
+ * - `subprocesses` subprocess reference when the workflow was invoked
+ * using a call activity. It can be define as one or multiple ones case
+ * the workflow was invoked from subprocess inside another subprocess.
+ * @returns mock for 'Check income request' completed task.
+ */
 export function createCheckIncomeRequestTaskMock(options?: {
   subprocesses?: WorkflowSubprocesses;
 }): Record<string, unknown> {
