+++ sources/packages/backend/workflow/test/test-utils/mock/cra-integration-income-verification/create-income-request-task-mock.ts
@@ -3,8 +3,20 @@ import {
   createMockedWorkerResult,
   WorkflowServiceTasks,
   WorkflowSubprocesses,
-} from "..";
+} from "../..";
 
+/**
+ * Create the mock for 'Create Income request' completed task
+ * and also publish the message to unblock the workflow.
+ * @param options mock options.
+ * - `incomeVerificationId` income verification that will be waiting for
+ * the message to be unblocked.
+ * - `subprocesses` subprocess reference when the workflow was invoked
+ * using a call activity. It can be define as one or multiple ones case
+ * the workflow was invoked from subprocess inside another subprocess.
+ * @returns mock for 'Create Income request' completed task and also publish
+ * the message to unblock the workflow.
+ */
 export function createIncomeRequestTaskMock(options?: {
   incomeVerificationId: number;
   subprocesses?: WorkflowSubprocesses;
