--- sources/packages/backend/workflow/test/test-utils/mock/verify-application-exceptions-task-mock.ts
@@ -1,15 +1,0 @@
-import { ApplicationExceptionStatus } from "@sims/sims-db";
-import { createMockedWorkerResult, WorkflowServiceTasks } from "..";
-
-export function createVerifyApplicationExceptionsTaskMock(options?: {
-  status: ApplicationExceptionStatus;
-}) {
-  const applicationExceptionStatus =
-    options?.status ?? ApplicationExceptionStatus.Approved;
-  return createMockedWorkerResult(
-    WorkflowServiceTasks.VerifyApplicationExceptions,
-    {
-      jobCompleteMock: { applicationExceptionStatus },
-    },
-  );
-}
